refactor: add code separation into modules
