+++ js/displayApp.js
@@ -1,0 +1,3 @@
+export default function displayApp(getMainContent) {
+  document.querySelector("body").insertAdjacentHTML("afterbegin", getMainContent);
+};
