+++ js/initLanguage.js
@@ -1,0 +1,7 @@
+export default function initLanguage(){
+  let language = localStorage.getItem("Lan");
+  if (language === null) {
+    localStorage.setItem("Lan", "en");
+    language = localStorage.getItem("Lan");
+  }
+}
