+++ js/state.js
@@ -1,0 +1,9 @@
+let state = {
+ capsLock : false,
+ shift : false,
+ control : false,
+ keyboardUP : false,
+ value : [],
+}
+
+export default state
